Properly check key size or bits type as int
The key_size or bits as passed to key generation functions
must of type int.

Previous code was raising an exception when the key size was
a string as a result of being a variable of a function argument.
Since variable values are not known when passed through functions,
it defaulted to a string of the name of the function arg.

Signed-off-by: Eric Brown <eric.brown@securesauce.dev>