Make the artifact argument of a Result to be lazy set (#271)
In order to simplify the rules, this change reduces the need to pass the
file artifact when initializing the Result. This amounts to less code in
each Rule. The artifact still gets set, but later in the parser. As a
result, the snippet and artifact URI are also set later.

Signed-off-by: Eric Brown <eric.brown@securesauce.dev>