Rework how a typed parameter is handled
This change utilizes the visitor pattern to process nodes of
type typed_parameter and typed_default_parameter instead of
manually checking children of the function def. This works because
typed parameters are only within function defs.

This commit also adds test cases to ensure this is working.

This change also protects against the traceback found when
encountering an unexpected *arg as a function parameter.

Fixes #310

Signed-off-by: Eric Brown <eric.brown@securesauce.dev>