Add security-severity property to sarif rules
In order for GitHub to render a High, Medium, Low for results
it requires the security-severity to be set.

This also means each rule needs a default level specific to it
to be set. Before it was defaulting to warning, but now will
default to the minimum possible value of level in results.

Signed-off-by: Eric Brown <eric.brown@securesauce.dev>