Fix so preignore properly ignores tests (#378)
During testing, it was found that adding a "tests" to .preignore did not
result in the ignoring of that directory in the repo.

It seems the path is needed as input to is_ignored() rather than just
file name.

Also noticed that .git files are getting scanned. Normally they should
be ignored as part of .gitignore, but its not typical to include there.
Seems Git automatically ignores that directory. So I added it as a
global_patterns to IgnoreFilterManager. Also figured that it would be
good to include global_ignore_file_paths as was in the example for this
class to exclude user defined ignores and repo defined ignores.

Signed-off-by: Eric Brown <eric.brown@securesauce.dev>