Ensure the tests import hashlib
There were a number of hmac tests that didn't import hashlib.
This should have caused precli to be unable to resolve it, but
because it is an attribute in this case instead of a call, it's
probably getting treated as a string which just works.

Also made sure files have no hyphens (md5-sha1)

Signed-off-by: Eric Brown <eric.brown@securesauce.dev>