Fix for cases of a reassignment of a variable (#392)
It is possible for a variable to be used multiple times and even with
multiple types prior to its use in the function call. This didn't work
because the left hand side of the assignment would be resolved into its
complete fully qualified function name.

However, it's also important to keep track of the identifier name in the
symbol table in case it is reused. That way the value is properly
overwritten.

Signed-off-by: Eric Brown <eric.brown@securesauce.dev>