Check for no action passed in argparse rule
The default value for action when no value is passed is "store"
which is the exact value we are trying to detect for issues. So
if a program creates CLI arguments via add_argument with api-key
or password arg and unset action, it needs to surface this as an
issue.

Signed-off-by: Eric Brown <eric.brown@securesauce.dev>