Do update check only if git target or gist output
This modifies the run of the CLI to only check for new PyPI updates
to the package if a given target is a GitHub address or the output
is designated to Gist. Both of these are remote, and so the cost
to check an update (also remote) only runs in the same case.

Signed-off-by: Eric Brown <eric.brown@securesauce.dev>