Add suggested fixes for the hashlib weak hash rule (#403)
The hashlib_weak_hash rule currently reports issues, but doesn't suggest
fixes. This change will now introduce fix suggestions. Namely it will
recommend using sha256.

This commit also changed the location of the reported issue to be the
actually function name or parameter so it is now more precise. This also
requires updates to the tests.

Signed-off-by: Eric Brown <eric.brown@securesauce.dev>