Handle non-UTF-8 encoding files
Python files default to be encoded in UTF-8, but they don't have
to be. At the head of Python file, you can define the encoding.

This change modifies the convenience function of the Node to return
a str object in the encoding of the file, honoring this header and
defaulting to utf-8.

See:
https://docs.python.org/3/howto/unicode.html#unicode-literals-in-python-source-code

Fixes: #468
Fixes: #473
Fixes: #472

Signed-off-by: Eric Brown <eric.brown@securesauce.dev>