Show download progress in file size
The Progress class of Rich supports showing progress as download
progress in bytes (MB, GB, etc). Ideally we would also show the
progress as a ratio to the total zipball file size, but unforutnately
the Content-Length header is not reliably returned.

The progress of the extraction was also modified to show progress
in terms of files extracted vs. the total files in the zip.

Signed-off-by: Eric Brown <eric.brown@securesauce.dev>