Store resolve dict values and store dicts in symtab
Add support for resolving values that are dicts. This involves
handling the defining of a dict and the reference of one with
the subscript node.

This change will also store dict values in the symbol table among
the others.

Also adds tests in hashlib that unit tests passing tuple and dicts.

Signed-off-by: Eric Brown <eric.brown@securesauce.dev>