Add new rule around loose permission in os module
Any usage of os.chmod and the like, the mode for the permissions
for a file should be checked to ensure the permissions are not
considered loose or incorrect. Incorrect in this case is using
world write, group write, world execute, group execute. Care
should be taken when using those.

Partially-resolves: #217

Signed-off-by: Eric Brown <eric.brown@securesauce.dev>