Add example of running command as animated gif
This change to the README uses asciinema to generate a cast
of running precli to analyze a test file and display the output.
The cast was then converted to an animated gif using agg.

asciinema rec example.cast
agg example.cast example.gif --font-dir=$HOME/Downloads/Noto_Emoji/static

Note: the font library was included so the emoji would render

https://github.com/asciinema/asciinema
https://github.com/asciinema/agg

Signed-off-by: Eric Brown <eric.brown@securesauce.dev>