Refactor how wildcard imports are processed
This change moves the wild card processing from the particular
parser into the rule class. The analyze_node() call mechanism
is being reused to do wildcard processing.

This is a strategy that could later be used to actually load
identifiers of an import into the symbol table as they are loaded.

Signed-off-by: Eric Brown <eric.brown@securesauce.dev>