Fix augmented assignment (#585)
It's possible that part of the expression of an augmented assignment
doesn't resolve to an int. In such a case, the current code tries to
bitwise compute the values anyway. This of course results in a syntax
error.

This fix in this change ensures both left and right side of the
expression resolve to an int value. Other bitwise computable values in
the future could also be added.

Signed-off-by: Eric Brown <eric.brown@securesauce.dev>