New Python rule checks for ssl use with no timeout (#609)
The ssl.get_server_certificate function has a timeout parameter that
defaults to the global timeout which itself defaults to None. When the
value is None, the socket blocks forever.

Note: the timeout parameter wasn't added till Python 3.10. To fix this
on earlier Python versions, someone would need to adjust the global
timeout value.

PY046

Signed-off-by: Eric Brown <eric.brown@securesauce.dev>