Enhance timeout rules ot honor the global timeout
Python code can use a function socket.setdefaulttimeout() to
set a global timeout value for all network classes functions
that use it. Therefore, the Precli rules that check for an
undefined timeout should honor this global value.

Note, currently this code will only recognize a global value of
the timeout on a per-file basis. Later an inter-file global
state of variables should be maintained, but this requires
better analyzing of the code in the expected order of execution.

Signed-off-by: Eric Brown <eric.brown@securesauce.dev>