Fix broken handling of file via stdin (#615)
The code was using pygments to guess the language of the stdin code
using function guess_lexer. Unfortunately, it doesn't work as expected.
It would not detect Go or Java reliably at all. The only reason it would
detect Python is because it would default to it in some way.

So the new approach is to attempt to parse the code using tree_sitter
and check for a syntax error. Of course, even this won't be 100%
accurate, but much better than the previous method.

Signed-off-by: Eric Brown <eric.brown@securesauce.dev>