Honor umask in calculating effective mode
Two Python loose permission rules will return results
when a function is used with a default permission mode
that is considered risky.

However, these rules were not accounting for the umask
that is typically set in the operating system and Python
honors when calculating the effective mode.

This change applies an assumed umask of 0o022, even though
it can vary to calculate effective modes for cases when
the default mode of a function is used.

If a mode value is passed to the function, it still uses
that mode argument as the mode to evaluate.

Signed-off-by: Eric Brown <eric.brown@securesauce.dev>