Ignore test code in analysis (#640)
This change, Python specific, determines whether code that is being
analyze is unit test or not. By default, the Python parser will not
execute rules for test code.

Test code is determined if the scope is within a function that starts
with "test_" or withing a "tests" directory or file name that starts
with "test_".

Currently the option to skip test code is on the Python parser so that
precli's own test code still works. But eventually this should become
part of some configuration options for precli.

Signed-off-by: Eric Brown <eric.brown@securesauce.dev>