Default the assert rule to be disabled (#641)
The point of the assert rule is to catch scenarios where Python builtin
assert is used to guard or as a conditional in functions and not realize
that running Python with optimizations (-O) would remove those assert
statements.

However, in practice, its unlikely folks use the optimize compile option
as it doesn't provide much "optimization" anyway. Frankly it does very
very little. Therefore this change defaults the assert rule to disabled
to cut down on noise and probably false positives.

Signed-off-by: Eric Brown <eric.brown@securesauce.dev>