Fix the output option
The output option was broken in such that argparse would create
the file if it didn't exist, then the code would additionally
check if the file exist, which it would.

This change removes the overwrite checking code and simply changes
argparse to error if the file already exists. It will be up to
the user to specify a different file or delete the existing.

Signed-off-by: Eric Brown <eric.brown@securesauce.dev>