Fix traceback when not passing file to -o option (#698)
If precli-init is run without an argument for the -o, it will fail with
a traceback due to the argparse.FileType not allowing a default of a str
for a file name.

This change does the logic of detecting and opening files manually on
its own. It will present an error if the file already exists to prevent
overwriting. Except in the case where the file is pyproject.toml and the
settings will be appended to it.

Signed-off-by: Eric Brown <eric.brown@securesauce.dev>