Ability to define custom rules via a YAML file
This change adds a new ability to create rules via a
simple YAML file instead of in pure Python code.

A YAML file defines some metadata and a query string.
The query string is used to do a tree-sitter query
and capture. If captures are found, the issue is added
as a result like normal.

Note: this method is not as accurate as the pure Python
rules as it doesn not track data flow. However, it could
be easier plugin model for stakeholders wishing to
write their own rules.

Closes: #769

Signed-off-by: Eric Brown <eric.brown@securesauce.dev>