Custom rule example with Ruby language
This is an example of using the custom rule function with
an unsupported language such as Ruby. The prerequiste is
the user needs to install tree_sitter_ruby for this to
work.

Note: this doesn't work for all tree-sitter parsers available
since the tree-sitter in use by precli requires language
version 15 to work. So C and Rust currently do not work,
possibly others.

Signed-off-by: Eric Brown <eric.brown@securesauce.dev>